requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
passlib[bcrypt]
bcrypt<4.0
python-jose[cryptography]
pydantic[email]
pydantic-settings
python-multipart
google-auth
pytest
